src/data-processor/requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.36.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.1
